# config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Новое в версии: - Добавили таблицу статусов TruckEvents, поле LastTruckEventID, LastTruckEventName в модель Waybills. - Добавили поля Departure - флаг выезда транспорта и DepartureDate - дата выезда транспорта. - Добавили новое поле в модель Waybills - Posted, обозначает проведение документа по учету, и возможность подписки по этому полю. - Добавили метод для помещения статусов /truckEvents/post. ",
        "version": "1.5.4",
        "title": "Bulkdata API - хранилище для товарно-транспортных накладных и лабораторных анализов.",
        "contact": {
            "email": "ineshyk@prncp.org"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/waybills": {
            "get": {
                "tags": [
                    "developers"
                ],
                "summary": "get all waybills to you account",
                "operationId": "getWaybills",
                "description": "By passing in the appropriate options, you can get\navailable updated waybills in the system\n",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "sourceID",
                        "description": "pass an optional source id for looking up current waybills",
                        "required": true,
                        "type": "string",
                        "format": "uid"
                    },
                    {
                        "in": "query",
                        "name": "id",
                        "description": "pass an optional id for looking up current waybill",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uid"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "maximum number of records to return",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0,
                        "maximum": 50
                    }
                ],
                "responses": {
                    "200": {
                        "description": "search results matching criteria",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Waybill"
                            }
                        }
                    },
                    "400": {
                        "description": "bad input parameter"
                    }
                }
            },
            "patch": {
                "tags": [
                    "developers"
                ],
                "summary": "set waybills inactive",
                "operationId": "pushWaybills",
                "description": "By passing in the appropriate options, you can set\nwaybills inactive\n",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "sourceID",
                        "description": "pass an optional source id for looking up current waybills",
                        "required": true,
                        "type": "string",
                        "format": "uid"
                    },
                    {
                        "in": "body",
                        "name": "waybillKey",
                        "description": "Waybill to incative",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/WaybillKey"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "search results matching criteria",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/WaybillKey"
                            }
                        }
                    },
                    "400": {
                        "description": "bad input parameter"
                    }
                }
            },
            "post": {
                "tags": [
                    "developers"
                ],
                "summary": "adds an waybill",
                "operationId": "addWaybill",
                "description": "Adds an waybill to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "waybill",
                        "description": "Waybill to add",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Waybill"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "waybill created"
                    },
                    "400": {
                        "description": "invalid input, object invalid"
                    },
                    "409": {
                        "description": "an existing waybill already exists"
                    }
                }
            },
            "put": {
                "tags": [
                    "developers"
                ],
                "summary": "delete an waybill",
                "operationId": "delWaybill",
                "description": "Delete an waybill from the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "delWaybill",
                        "description": "Waybill to delete",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/WaybillKey"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "waybill deleted"
                    },
                    "400": {
                        "description": "invalid input, object invalid"
                    }
                }
            }
        },
        "/labAnalysis": {
            "get": {
                "tags": [
                    "developers"
                ],
                "summary": "get all lab analysis to you account",
                "operationId": "getlabAnalysis",
                "description": "By passing in the appropriate options, you can get\navailable updated laboratory abalysis in the system\n",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "sourceID",
                        "description": "pass an optional source id for looking up current laboratory analysis",
                        "required": true,
                        "type": "string",
                        "format": "uid"
                    },
                    {
                        "in": "query",
                        "name": "id",
                        "description": "pass an optional id for looking up current lab analys",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uid"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "maximum number of records to return",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0,
                        "maximum": 50
                    }
                ],
                "responses": {
                    "200": {
                        "description": "search results matching criteria",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/LabAnalys"
                            }
                        }
                    },
                    "400": {
                        "description": "bad input parameter"
                    }
                }
            },
            "patch": {
                "tags": [
                    "developers"
                ],
                "summary": "set waybills inactive",
                "operationId": "pushLabAnalysis",
                "description": "By passing in the appropriate options, you can set\nlab analys inactive\n",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "sourceID",
                        "description": "pass an optional source id for looking up current laboratory analysis",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "labAnalysKey",
                        "description": "Laboratory analys to incative",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/LabAnalysKey"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "search results matching criteria",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/LabAnalysKey"
                            }
                        }
                    },
                    "400": {
                        "description": "bad input parameter"
                    }
                }
            },
            "post": {
                "tags": [
                    "developers"
                ],
                "summary": "adds an waybill",
                "operationId": "addLabAnalysis",
                "description": "Adds an laboratory analysis to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "labAnalys",
                        "description": "Laboratory analysis to add",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/LabAnalys"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "lab analysis created"
                    },
                    "400": {
                        "description": "invalid input, object invalid"
                    },
                    "409": {
                        "description": "an existing analysis already exists"
                    }
                }
            },
            "put": {
                "tags": [
                    "developers"
                ],
                "summary": "delete an lab analys",
                "operationId": "delLabAnalys",
                "description": "Delete an laboratory analys from the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "delLabAnalys",
                        "description": "Laboratory analys to delete",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/LabAnalysKey"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "laboratory analys deleted"
                    },
                    "400": {
                        "description": "invalid input, object invalid"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "WaybillKey": {
            "type": "object",
            "required": [
                "EntryGUID"
            ],
            "properties": {
                "EntryGUID": {
                    "type": "string",
                    "format": "uuid",
                    "uniqueItems": true,
                    "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
                }
            }
        },
...
        "LabAnalysKey": {
            "type": "object",
            "required": [
                "FormID"
            ],
            "properties": {
                "FormID": {
                    "type": "string",
                    "format": "uuid",
                    "uniqueItems": true,
                    "example": "z256561ee-6c54-4b01-90e6-d701748f0456"
                }
            }
        },
        "LabAnalys": {
            "type": "object",
            "required": [
                "FormID",
                "Type",
                "TransportShipmentID",
                "Active",
                "ConsigneeRegCode"
            ],
            "properties": {
                "FormID": {
                    "type": "string",
                    "format": "uuid",
                    "uniqueItems": true,
                    "example": "z256561ee-6c54-4b01-90e6-d701748f0456",
                    "description": "Уникальный идентификатор лабораторного анализа."
                },
                "Type": {
                    "type": "string",
                    "maxLength": 15,
                    "example": "Полный.",
                    "description": "Тип лабораторного анализа (полный, расчетный, экспресс)."
                },
                "TransportShipmentID": {
                    "type": "string",
                    "format": "uuid",
                    "uniqueItems": true,
                    "example": "d290f1ee-6c54-4b01-90e6-d701748f0851",
                    "description": "Уникальный идентификатор товарной операции."
                },
                "Active": {
                    "type": "boolean",
                    "description": "Если какие-либо данные удаляются у источника данных, устанавливается = нет. 0=удален."
                },
                "OwnerStateRegCode": {
                    "type": "string",
                    "maxLength": 12,
                    "example": "31333767",
                    "description": "Код едрпоу сосбтвенника (ИНН)."
                },
                "ConsigneeRegCode": {
                    "type": "string",
                    "maxLength": 15,
                    "example": "31333767",
                    "description": "Код едрпоу предприятия получателя (ИНН)."
                },
                "ConsignorRegCode": {
                    "type": "string",
                    "maxLength": 15,
                    "example": "25255267",
                    "description": "Код едрпоу предприятия отправителя (ИНН)."
                },
                "No": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 20,
                    "example": "ЭЛ01-0004534566",
                    "description": "Номер анализной карточки."
                },
                "Date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2017-03-15T18:25:43.511Z"
                },
                "ScanCode": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 20,
                    "example": "0123456789123",
                    "description": "Дата анализной карточки"
                },
                "HarvestYear": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2017-01-01T00:00:00.000Z",
                    "description": "Год урожая."
                },
                "WaybillNo": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 30,
                    "example": "ЭЛ01-20170315-001",
                    "description": "Номер товарной операции, к которой подвязан документ."
                },
                "WaybillDate": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2017-03-14T14:40:43.511Z",
                    "description": "Дата товарной операции, к которой подвязан документ."
                },
                "Certificate1No": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 40,
                    "example": "12345678910",
                    "description": "Номер сертификата протокола безопасности."
                },
                "Certificate1Date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2017-03-10T14:40:43.511Z",
                    "description": "Дата сертификата протокола безопасности."
                },
                "Certificate2No": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 40,
                    "example": "0123456789012",
                    "description": "Номер сертификата протокола ГМО."
                },
                "Certificate2Date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2017-03-09T14:40:43.511Z",
                    "description": "Дата сертификата протокола ГМО."
                },
                "Certificate3No": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 40,
                    "example": "12345678910",
                    "description": "Номер сертификата качества."
                },
                "Certificate3Date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2017-03-10T14:40:43.511Z",
                    "description": "Дата сертификата качества."
                },
                "EntryType": {
                    "type": "boolean",
                    "description": "Вид движения 0-приход, 1 - расход"
                },
                "ItemGUID": {
                    "type": "string",
                    "maxLength": 56,
                    "example": "КУК3КЛ",
                    "description": "Индетификатотр груза."
                },
                "ItemName": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "Кукурудза 3 класу кормові потреби",
                    "description": "Наименование груза."
                },
                "ItemNavCode": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "NAV5",
                    "description": "Код груза из классификатора."
                },
                "LabOperator": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "Васильева А.П.",
                    "description": "ФИО Лаборанта, который провел анализ"
                },
                "OwnerType": {
                    "type": "integer",
                    "format": "int32",
                    "maxLength": 1,
                    "example": 0,
                    "description": "Тип владельца груза (0 - сторонние, 1 - Кернел, 2- ДАПР."
                },
                "ConsigneeType": {
                    "type": "integer",
                    "format": "int32",
                    "maxLength": 1,
                    "example": 0,
                    "description": "Тип получателя (0 - сторонние, 1 - Кернел, 2- ДАПР."
                },
                "ConsignorType": {
                    "type": "integer",
                    "format": "int32",
                    "maxLength": 1,
                    "example": 0,
                    "description": "Тип отправителя (0 - сторонние, 1 - Кернел, 2- ДАПР."
                },
                "Qualities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/LabAnalysLine"
                    },
                    "description": "Таблица с качественными показателями."
                }
            }
        },
        "LabAnalysLine": {
            "type": "object",
            "required": [
                "FormID",
                "AnalisysCode",
                "AnalisysName",
                "ValueType"
            ],
            "properties": {
                "FormID": {
                    "type": "string",
                    "format": "uuid",
                    "uniqueItems": true,
                    "example": "z256561ee-6c54-4b01-90e6-d701748f0456",
                    "description": "Уникальный идентификатор лабораторного анализа."
                },
                "AnalisysCode": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 50,
                    "example": "МинеральнаяПримесь",
                    "description": "Идентификатор качественных показателей. "
                },
                "AnalisysName": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 100,
                    "example": "Минеральная примесь",
                    "description": "Наименование качественного показателя."
                },
                "ValueType": {
                    "type": "boolean",
                    "description": "Тип значения (0-число, 1-строка)."
                },
                "UOM": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 100,
                    "example": "%",
                    "description": "Единица измерения."
                },
                "Value": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 40,
                    "example": "",
                    "description": "Строковое значение показателя."
                },
                "Value2": {
                    "type": "number",
                    "format": "double",
                    "example": 2,
                    "description": "Числовое знаениче показателя."
                },
                "ParentAnalisysCode": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 50,
                    "example": "СорнаяПримесь",
                    "description": "Уникальный идентификатор родителя качественного показателя."
                },
                "Inclusion": {
                    "type": "boolean",
                    "description": "Признак влияния значения текущего показателя на итоговый показатель владельца."
                },
                "ShowinLabes": {
                    "type": "boolean",
                    "description": "Признак отображения текущего показателя в печатной форме документа."
                }
            }
        },
...
    }
}